Misc changes and support for disambiguation

+++ TextConverter/src/main/kotlin/TextConverterFeatureSummaryGenerator.kt
@@ -72,6 +72,7 @@ object TextConverterFeatureSummaryGenerator: TextConverterProcessorBase()
   /****************************************************************************/
   override fun pre (): Boolean
   {
+    if (C_InputType != InputType.USX) return true
     deleteFile(Pair(StandardFileLocations.getTextFeaturesFilePath(), null))
     return true
   }
@@ -80,7 +81,7 @@ object TextConverterFeatureSummaryGenerator: TextConverterProcessorBase()
   /****************************************************************************/
   override fun runMe (): Boolean
   {
-    return true
+    return C_InputType == InputType.USX
   }
